modified to make creating sparce boundary matrix more efficient https://github.com/sriraam-ap/TD4-TopologicalPersistence/issues/7, organized dir structures, added test code for small tesecases
