+++ README.md
@@ -5,3 +5,17 @@ Emails for project submission:
 — steve.oudot@inria.fr
 — ringyi.li@polytechnique.edu
 — julie.mordacq@inria.fr
+
+# Environment
+- python 3.11.2 or later
+
+# tests
+Run following code on the project root directory after installing pytest.
+
+```
+kazetof@inspiron:~/Works/IPParis/TDA/TD4/TD4-TopologicalPersistence$ pwd
+/home/kazetof/Works/IPParis/TDA/TD4/TD4-TopologicalPersistence
+kazetof@inspiron:~/Works/IPParis/TDA/TD4/TD4-TopologicalPersistence$ ls
+LICENSE  README.md  __pycache__  examples  filtrations  persistence  requirements.txt  tests
+kazetof@inspiron:~/Works/IPParis/TDA/TD4/TD4-TopologicalPersistence$ PYTHONPATH=./ pytest ./tests -s
+```
